# scratch/schema-openapi.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "title": "standard public schema",
    "version": "14.5"
  },
  "host": "rmmaglalrdbqvvbyhnpx.supabase.co:443",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "produces": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "paths": {
...
    "/entries": {
      "get": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.entries.id"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.user_id"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.content"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.word_count"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.session_id"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.updated_at"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.client_id"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.cycle_id"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.cycle_day"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.reflection_text_encrypted"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.reflection_text_iv"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.new_entry_text_encrypted"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.new_entry_text_iv"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.entry_type"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.written_at"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.synced_at"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.mode"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.thread_response"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.open_thread_id"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.exercise_id"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.reflection_ei"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.reflection_pr"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.reflection_sa"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.new_entry_ei"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.new_entry_pr"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.new_entry_sa"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.day_ei"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.day_pr"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.day_sa"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.confidence_flag"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.confidence_reason"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.scoring_status"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.crisis_flag"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.crisis_type"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.crisis_flagged_at"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.reflection_suppressed"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.risk_language_quote"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.crisis_checked"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.arc_scoring_applied"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.arc_scoring_note"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/order"
          },
          {
            "$ref": "#/parameters/range"
          },
          {
            "$ref": "#/parameters/rangeUnit"
          },
          {
            "$ref": "#/parameters/offset"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/preferCount"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/entries"
              },
              "type": "array"
            }
          },
          "206": {
            "description": "Partial Content"
          }
        },
        "tags": [
          "entries"
        ]
      },
      "post": {
        "parameters": [
          {
            "$ref": "#/parameters/body.entries"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/preferPost"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "tags": [
          "entries"
        ]
      },
      "delete": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.entries.id"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.user_id"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.content"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.word_count"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.session_id"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.updated_at"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.client_id"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.cycle_id"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.cycle_day"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.reflection_text_encrypted"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.reflection_text_iv"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.new_entry_text_encrypted"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.new_entry_text_iv"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.entry_type"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.written_at"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.synced_at"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.mode"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.thread_response"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.open_thread_id"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.exercise_id"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.reflection_ei"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.reflection_pr"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.reflection_sa"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.new_entry_ei"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.new_entry_pr"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.new_entry_sa"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.day_ei"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.day_pr"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.day_sa"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.confidence_flag"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.confidence_reason"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.scoring_status"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.crisis_flag"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.crisis_type"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.crisis_flagged_at"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.reflection_suppressed"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.risk_language_quote"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.crisis_checked"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.arc_scoring_applied"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.arc_scoring_note"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "entries"
        ]
      },
      "patch": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.entries.id"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.user_id"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.content"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.word_count"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.session_id"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.updated_at"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.client_id"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.cycle_id"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.cycle_day"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.reflection_text_encrypted"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.reflection_text_iv"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.new_entry_text_encrypted"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.new_entry_text_iv"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.entry_type"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.written_at"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.synced_at"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.mode"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.thread_response"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.open_thread_id"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.exercise_id"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.reflection_ei"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.reflection_pr"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.reflection_sa"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.new_entry_ei"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.new_entry_pr"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.new_entry_sa"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.day_ei"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.day_pr"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.day_sa"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.confidence_flag"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.confidence_reason"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.scoring_status"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.crisis_flag"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.crisis_type"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.crisis_flagged_at"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.reflection_suppressed"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.risk_language_quote"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.crisis_checked"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.arc_scoring_applied"
          },
          {
            "$ref": "#/parameters/rowFilter.entries.arc_scoring_note"
          },
          {
            "$ref": "#/parameters/body.entries"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "entries"
        ]
      }
    },
...
  },
...
  "parameters": {
...
    "preferReturn": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferCount": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "count=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferPost": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none",
        "resolution=ignore-duplicates",
        "resolution=merge-duplicates"
      ],
      "in": "header",
      "type": "string"
    },
    "select": {
      "name": "select",
      "description": "Filtering Columns",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "order": {
      "name": "order",
      "description": "Ordering",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "range": {
      "name": "Range",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "header",
      "type": "string"
    },
    "rangeUnit": {
      "name": "Range-Unit",
      "description": "Limiting and Pagination",
      "required": false,
      "default": "items",
      "in": "header",
      "type": "string"
    },
    "offset": {
      "name": "offset",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "limit": {
      "name": "limit",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "body.entries": {
      "name": "entries",
      "description": "entries",
      "required": false,
      "in": "body",
      "schema": {
        "$ref": "#/definitions/entries"
      }
    },
    "rowFilter.entries.id": {
      "name": "id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.user_id": {
      "name": "user_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.content": {
      "name": "content",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.word_count": {
      "name": "word_count",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.created_at": {
      "name": "created_at",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.session_id": {
      "name": "session_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.updated_at": {
      "name": "updated_at",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.client_id": {
      "name": "client_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.cycle_id": {
      "name": "cycle_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.cycle_day": {
      "name": "cycle_day",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.reflection_text_encrypted": {
      "name": "reflection_text_encrypted",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.reflection_text_iv": {
      "name": "reflection_text_iv",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.new_entry_text_encrypted": {
      "name": "new_entry_text_encrypted",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.new_entry_text_iv": {
      "name": "new_entry_text_iv",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.entry_type": {
      "name": "entry_type",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.written_at": {
      "name": "written_at",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.synced_at": {
      "name": "synced_at",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.mode": {
      "name": "mode",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.thread_response": {
      "name": "thread_response",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.open_thread_id": {
      "name": "open_thread_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.exercise_id": {
      "name": "exercise_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.reflection_ei": {
      "name": "reflection_ei",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.reflection_pr": {
      "name": "reflection_pr",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.reflection_sa": {
      "name": "reflection_sa",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.new_entry_ei": {
      "name": "new_entry_ei",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.new_entry_pr": {
      "name": "new_entry_pr",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.new_entry_sa": {
      "name": "new_entry_sa",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.day_ei": {
      "name": "day_ei",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.day_pr": {
      "name": "day_pr",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.day_sa": {
      "name": "day_sa",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.confidence_flag": {
      "name": "confidence_flag",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.confidence_reason": {
      "name": "confidence_reason",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.scoring_status": {
      "name": "scoring_status",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.crisis_flag": {
      "name": "crisis_flag",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.crisis_type": {
      "name": "crisis_type",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.crisis_flagged_at": {
      "name": "crisis_flagged_at",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.reflection_suppressed": {
      "name": "reflection_suppressed",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.risk_language_quote": {
      "name": "risk_language_quote",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.crisis_checked": {
      "name": "crisis_checked",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.arc_scoring_applied": {
      "name": "arc_scoring_applied",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.entries.arc_scoring_note": {
      "name": "arc_scoring_note",
      "description": "Details of arc scoring values when applied",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
  },
...
}