# src/app/api/entries/route.ts
import { NextRequest, NextResponse } from 'next/server';
import { supabase } from '../../../lib/db';
import { getAuthenticatedUser } from '../../../lib/auth-helper';
import { triggerAIProcessing, checkWeeklyAndMonthlySummary } from '../../../lib/queue/triggers';

/**
 * GET /api/entries: Fetches all journal entries for the user from Supabase.
 */
export async function GET(request: NextRequest) {
  try {
    const authUser = await getAuthenticatedUser(request);
    if (!authUser) {
      return NextResponse.json(
        { error: { code: 'AUTH_REQUIRED', message: 'Authentication is required.' } },
        { status: 401 }
      );
    }

    const search = request.nextUrl.searchParams.get('search');
    const cycleId = request.nextUrl.searchParams.get('cycleId');
    const classification = request.nextUrl.searchParams.get('classification');
    const isCrisis = request.nextUrl.searchParams.get('isCrisis');
    const startDate = request.nextUrl.searchParams.get('startDate');
    const endDate = request.nextUrl.searchParams.get('endDate');

    let query;
    let fallbackQuery;

    if (classification) {
      query = supabase
        .from('entries')
        .select('*, reflections!inner(*), daily_sessions!fk_daily_sessions_entry(day_number), cycles(*)')
        .eq('user_id', authUser.userId)
        .eq('reflections.classification', classification);

      fallbackQuery = supabase
        .from('entries')
        .select('*, reflections!inner(*), cycles(*)')
        .eq('user_id', authUser.userId)
        .eq('reflections.classification', classification);
    } else {
      query = supabase
        .from('entries')
        .select('*, reflections(*), daily_sessions!fk_daily_sessions_entry(day_number), cycles(*)')
        .eq('user_id', authUser.userId);

      fallbackQuery = supabase
        .from('entries')
        .select('*, reflections(*), cycles(*)')
        .eq('user_id', authUser.userId);
    }

    if (search) {
      query = query.ilike('content', `%${search}%`);
      fallbackQuery = fallbackQuery.ilike('content', `%${search}%`);
    }
    if (cycleId) {
      query = query.eq('cycle_id', cycleId);
      fallbackQuery = fallbackQuery.eq('cycle_id', cycleId);
    }
    if (isCrisis) {
      query = query.eq('crisis_flag', isCrisis === 'true');
      fallbackQuery = fallbackQuery.eq('crisis_flag', isCrisis === 'true');
    }
    if (startDate) {
      query = query.gte('created_at', startDate);
      fallbackQuery = fallbackQuery.gte('created_at', startDate);
    }
    if (endDate) {
      query = query.lte('created_at', endDate);
      fallbackQuery = fallbackQuery.lte('created_at', endDate);
    }

    query = query.order('created_at', { ascending: false });
    fallbackQuery = fallbackQuery.order('created_at', { ascending: false });

    let { data: entries, error } = await query;

    // Fallback if the database schema has relational join ambiguity
    if (error && (error.code === 'PGRST201' || error.code === 'PGRST200' || error.message.includes('relationship') || error.message.includes('column'))) {
      console.warn('[api/entries] Join query failed, falling back to simple select (schema may need migration):', error.message);
      const { data: simpleEntries, error: simpleError } = await fallbackQuery;
        
      if (simpleError) {
        console.error('Failed to fetch journal entries on fallback:', simpleError);
        return NextResponse.json(
          { error: { code: 'DATABASE_ERROR', message: 'Failed to retrieve journal entries.' } },
          { status: 500 }
        );
      }
      entries = simpleEntries;
    } else if (error) {
      console.error('Failed to fetch journal entries:', error);
      return NextResponse.json(
        { error: { code: 'DATABASE_ERROR', message: 'Failed to retrieve journal entries.' } },
        { status: 500 }
      );
    }

    // Fetch cycle vocab words to map vocab elements per entry cycle
    const { data: vocabRes } = await supabase
      .from('vocab_words')
      .select('cycle_id, word')
      .eq('user_id', authUser.userId);

    const formattedEntries = (entries || []).map((entry: any) => {
      // Normalize reflections (can be array or single object depending on PostgREST cardinality detection)
      const rawReflection = entry.reflections;
      const reflection = Array.isArray(rawReflection)
        ? (rawReflection[0] || null)
        : (rawReflection || null);
      delete entry.reflections;

      // Normalize daily_sessions (can be array or single object)
      const rawSession = entry.daily_sessions;
      const session = Array.isArray(rawSession)
        ? (rawSession[0] || null)
        : (rawSession || null);
      entry.daily_sessions = session;

      // Normalize cycles (can be array or single object)
      const rawCycle = entry.cycles;
      const cycle = Array.isArray(rawCycle)
        ? (rawCycle[0] || null)
        : (rawCycle || null);
      
      const cycleNum = cycle 
        ? (cycle.cycle_number !== undefined ? cycle.cycle_number : cycle.number)
        : null;
      delete entry.cycles;

      const cycleVocab = vocabRes
        ? vocabRes.filter((v: any) => v.cycle_id === entry.cycle_id).map((v: any) => v.word)
        : [];

      return {
        ...entry,
        cycle_number: cycleNum,
        reflection: reflection ? {
          ...reflection,
          vocabulary: cycleVocab
        } : null
      };
    });

    return NextResponse.json({
      success: true,
      entries: formattedEntries
    });

  } catch (error) {
    console.error('Entries GET Route Error:', error);
    return NextResponse.json(
      { error: { code: 'INTERNAL_ERROR', message: 'An unexpected server error occurred.' } },
      { status: 500 }
    );
  }
}

/**
 * POST /api/entries: Saves a new free-form journal entry in Supabase.
 */
export async function POST(request: NextRequest) {
  try {
    const authUser = await getAuthenticatedUser(request);
    if (!authUser) {
      return NextResponse.json(
        { error: { code: 'AUTH_REQUIRED', message: 'Authentication is required.' } },
        { status: 401 }
      );
    }

    const body = await request.json().catch(() => ({}));
    const { content } = body;

    if (!content || !content.trim()) {
      return NextResponse.json(
        { error: { code: 'BAD_REQUEST', message: 'Journal entry content cannot be empty.' } },
        { status: 400 }
      );
    }

    // Daily limit validation: Check if they have already written today
    const clientTodayStart = request.headers.get('x-client-today-start');
    const fallbackStart = new Date();
    fallbackStart.setUTCHours(0, 0, 0, 0);
    const startRange = clientTodayStart || fallbackStart.toISOString();

    const { data: existingEntries, error: checkError } = await supabase
      .from('entries')
      .select('id')
      .eq('user_id', authUser.userId)
      .gte('created_at', startRange)
      .limit(1);

    if (checkError) {
      console.error('Failed to check existing entries for daily limit:', checkError);
    } else if (existingEntries && existingEntries.length > 0) {
      return NextResponse.json(
        { error: { code: 'LIMIT_EXCEEDED', message: 'You have already completed your writing for today. The limit resets at midnight.' } },
        { status: 400 }
      );
    }

    const wordCount = content.trim().split(/\s+/).filter(Boolean).length;

    // Gating check: Block new entries if the latest cycle is completed but assessment is pending
    const { data: latestCycle } = await supabase
      .from('cycles')
      .select('*')
      .eq('user_id', authUser.userId)
      .order('created_at', { ascending: false })
      .limit(1)
      .maybeSingle();

    if (latestCycle && (latestCycle.status === 'COMPLETED' || latestCycle.status === 'complete') && !latestCycle.assessment_completed) {
      return NextResponse.json(
        { error: { code: 'ASSESSMENT_REQUIRED', message: 'You must complete your cycle assessment before writing new entries.' } },
        { status: 400 }
      );
    }

    // Fetch the active cycle for the user to populate cycle_id and cycle_day
    const { data: activeCycle } = await supabase
      .from('cycles')
      .select('id, start_date, total_days')
      .eq('user_id', authUser.userId)
      .in('status', ['ACTIVE', 'active'])
      .maybeSingle();

    const cycleId = activeCycle?.id || null;
    let cycleDay = 1;
    if (activeCycle) {
      const started = new Date(activeCycle.start_date);
      const today = new Date();
      const startMidnight = new Date(started.getFullYear(), started.getMonth(), started.getDate());
      const todayMidnight = new Date(today.getFullYear(), today.getMonth(), today.getDate());
      const diffTime = todayMidnight.getTime() - startMidnight.getTime();
      const calculatedDay = Math.floor(diffTime / (24 * 60 * 60 * 1000)) + 1;
      cycleDay = Math.min(activeCycle.total_days || 30, Math.max(1, calculatedDay));
    }

    const insertPayload: any = {
      user_id: authUser.userId,
      content: content.trim(),
      new_entry_text_encrypted: content.trim(), // for future AI workflow compatibility
      entry_type: 'new_only',
      word_count: wordCount,
      session_id: null,
      cycle_id: cycleId,
      cycle_day: cycleDay,
      written_at: new Date().toISOString()
    };

    let { data: newEntry, error } = await supabase
      .from('entries')
      .insert(insertPayload)
      .select()
      .single();

    // Fallback if the database schema is not fully migrated (missing new columns)
    if (error && (error.message.includes('column') || error.code === 'PGRST200' || error.code === '42703')) {
      console.warn('[api/entries] Insert failed with cycle/encryption columns, retrying with basic columns...');
      const fallbackPayload = {
        user_id: authUser.userId,
        content: content.trim(),
        word_count: wordCount,
        session_id: null
      };
      const { data: fallbackEntry, error: fallbackError } = await supabase
        .from('entries')
        .insert(fallbackPayload)
        .select()
        .single();
        
      if (fallbackError) {
        console.error('Failed to insert journal entry on fallback:', fallbackError);
        return NextResponse.json(
          { error: { code: 'DATABASE_ERROR', message: 'Failed to save journal entry.' } },
          { status: 500 }
        );
      }
      newEntry = fallbackEntry;
    } else if (error) {
      console.error('Failed to insert journal entry:', error);
      return NextResponse.json(
        { error: { code: 'DATABASE_ERROR', message: 'Failed to save journal entry.' } },
        { status: 500 }
      );
    }

    if (newEntry) {
      // Trigger background AI tasks and check weekly/monthly milestones
      triggerAIProcessing(newEntry.id, authUser.userId);
      checkWeeklyAndMonthlySummary(authUser.userId, newEntry.cycle_id, newEntry.cycle_day);
    }

    return NextResponse.json({
      success: true,
      entry: newEntry
    });

  } catch (error) {
    console.error('Entries POST Route Error:', error);
    return NextResponse.json(
      { error: { code: 'INTERNAL_ERROR', message: 'An unexpected server error occurred.' } },
      { status: 500 }
    );
  }
}